mv setup_metadb fixt to unit dir
